refactor: move scenario path to new()
